requirements.txt
django>=5.0
django-tenants
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-crontab
uuid
python-dotenv
python-dateutil